app/requirements.txt
dotenv==0.9.9
fastapi==0.116.1
httpx==0.28.1
langchain==0.3.27
langchain-community==0.3.29
langchain_core==0.3.76
langchain_google_genai==2.1.8
langchain_openai==0.3.28
pillow==11.3.0
pydantic==2.11.7
pydantic_settings==2.10.1
python-dotenv==1.1.1
starlette==0.47.2
tenacity==9.1.2
uvicorn==0.35.0